+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+### Added
+- Google Analytics improvements.
 
 ## [5.5.10] - 2026-02-05
 ### Fixed